Moves DomainAliasTestTrait.
