+++ domain_alias/tests/src/Functional/DomainAliasListBuilderTest.php
@@ -3,6 +3,7 @@
 namespace Drupal\Tests\domain_alias\Functional;
 
 use Drupal\Tests\domain\Functional\DomainTestBase;
+use Drupal\Tests\domain_alias\Traits\DomainAliasTestTrait;
 
 /**
  * Tests behavior for the domain list builder.
